requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
Jinja2==3.1.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==1.26.4
packaging==24.1
pandas==2.0.1
pillow==10.4.0
psutil==5.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-nmap==0.6.1
pytz==2024.1
scapy==2.4.5
six==1.16.0
tzdata==2024.1
